refactor database files
