+++ client/run_client.sh
@@ -1,2 +1,2 @@
 #!/bin/bash
-gcc -o client ./*.c && ./client
+gcc -o client ./*.c ../lib/*.c `mysql_config --cflags --libs` && ./client
