+++ server/run_server.sh
@@ -1,2 +1,2 @@
 #!/bin/bash
-gcc -o server ./*.c ./server_lib/*.c `mysql_config --cflags --libs` && ./server
+gcc -o server ./*.c ./server_lib/*.c ../lib/*.c `mysql_config --cflags --libs` && ./server
